+++ packages/vite-plugin/CHANGELOG.md
@@ -1,5 +1,26 @@
 # @vanilla-extract/vite-plugin
 
+## 5.0.3
+
+### Patch Changes
+
+- [#1592](https://github.com/vanilla-extract-css/vanilla-extract/pull/1592) [`35d6b3f`](https://github.com/vanilla-extract-css/vanilla-extract/commit/35d6b3fe5d297dc76f1629a9f54f21d0f97e7c9f) Thanks [@askoufis](https://github.com/askoufis)! - Disable the compiler's file watcher during builds
+
+- Updated dependencies [[`35d6b3f`](https://github.com/vanilla-extract-css/vanilla-extract/commit/35d6b3fe5d297dc76f1629a9f54f21d0f97e7c9f)]:
+  - @vanilla-extract/compiler@0.2.0
+
+## 5.0.2
+
+### Patch Changes
+
+- [#1583](https://github.com/vanilla-extract-css/vanilla-extract/pull/1583) [`6488e28`](https://github.com/vanilla-extract-css/vanilla-extract/commit/6488e28337106582231b418778c884f26e393219) Thanks [@askoufis](https://github.com/askoufis)! - Ignore Vite `base` config
+
+  Fixes a bug where imported asset URLs would be transformed by the compiler when setting the `base` option in Vite, resulting in incorrect resolution and bundling of these assets during the build process.
+
+- Updated dependencies [[`6488e28`](https://github.com/vanilla-extract-css/vanilla-extract/commit/6488e28337106582231b418778c884f26e393219)]:
+  - @vanilla-extract/compiler@0.1.3
+  - @vanilla-extract/integration@8.0.2
+
 ## 5.0.1
 
 ### Patch Changes